+++ lib/cellular_automata.dart
@@ -4,6 +4,7 @@ library cellular_automata;
 export 'package:cellular_automata/src/simulator.dart';
 export 'package:cellular_automata/src/cell_world.dart';
 export 'package:cellular_automata/src/rules/_ca_rules.dart';
+export 'package:cellular_automata/src/util/array_2d.dart';
 
 export 'package:cellular_automata/src/generators/_ca_generator.dart';
 export 'package:cellular_automata/src/generators/mathematical.dart';